# Generalized summations