--- a/tests/Tests/Sum.hs
+++ b/tests/Tests/Sum.hs
@@ -1,8 +1,14 @@
+{-# LANGUAGE FlexibleContexts #-}
+{-# LANGUAGE RankNTypes #-}
+{-# LANGUAGE ScopedTypeVariables #-}
+{-# LANGUAGE TypeApplications #-}
 {-# OPTIONS_GHC -fno-warn-orphans #-}
 
 module Tests.Sum (tests) where
 
 import Control.Applicative ((<$>))
+import Data.Functor.Identity
+import Foreign.C.Types
 import Numeric.Sum as Sum
 import Prelude hiding (sum)
 import Test.Tasty (TestTree, testGroup)
@@ -10,20 +16,20 @@ import Test.Tasty.QuickCheck (testProperty)
 import Test.QuickCheck (Arbitrary(..))
 import qualified Prelude
 
-t_sum :: ([Double] -> Double) -> [Double] -> Bool
+t_sum :: RealFloat a => ([a] -> a) -> [a] -> Bool
 t_sum f xs = f xs == trueSum xs
 
-t_sum_error :: ([Double] -> Double) -> [Double] -> Bool
+t_sum_error :: RealFloat a => ([a] -> a) -> [a] -> Bool
 t_sum_error f xs = abs (ts - f xs) <= abs (ts - Prelude.sum xs)
   where ts = trueSum xs
 
-t_sum_shifted :: ([Double] -> Double) -> [Double] -> Bool
+t_sum_shifted :: RealFloat a => ([a] -> a) -> [a] -> Bool
 t_sum_shifted f = t_sum_error f . zipWith (+) badvec
 
 trueSum :: (Fractional b, Real a) => [a] -> b
 trueSum xs = fromRational . Prelude.sum . map toRational $ xs
 
-badvec :: [Double]
+badvec :: RealFloat a => [a]
 badvec = cycle [1,1e16,-1e16]
 
 tests :: TestTree
@@ -37,51 +43,71 @@ tests = testGroup "Summation" [
       -- testProperty "t_sum_shifted" $ t_sum_shifted (sum id)
     ]
   , testGroup "Kahan" [
+      testGroup "Float " [
       -- tests that cannot pass:
-      -- testProprty "t_sum" $ t_sum (sum kahan)
-      -- testProperty "t_sum_error" $ t_sum_error (sum kahan)
+      -- testProperty "t_sum" $ t_sum @Float (sum kahan)
+      -- testProperty "t_sum_error" $ t_sum_error @Float (sum kahan)
 
       -- kahan summation only beats normal summation with large values
-      testProperty "t_sum_shifted" $ t_sum_shifted (sum kahan)
+        testProperty "t_sum_shifted" $ t_sum_shifted @Float (sum kahan)
+      ]
+    , testGroup "Double" [
+        testProperty "t_sum_shifted" $ t_sum_shifted @Double (sum kahan)
+      ]
     ]
-  , testGroup "KBN" [
-      testProperty "t_sum" $ t_sum (sum kbn)
-    , testProperty "t_sum_error" $ t_sum_error (sum kbn)
-    , testProperty "t_sum_shifted" $ t_sum_shifted (sum kbn)
-    ]
-  , testGroup "KB2" [
-      testProperty "t_sum" $ t_sum (sum kb2)
-    , testProperty "t_sum_error" $ t_sum_error (sum kb2)
-    , testProperty "t_sum_shifted" $ t_sum_shifted (sum kb2)
-    ]
+  , testGroup "KBN" $ testSum kbn
+  , testGroup "KB2" $ testSum kb2
   ]
 
-instance Arbitrary KahanSum where
+testSum :: ( Summation s Float
+           , Summation s Double
+           , Summation s CFloat
+           , Summation s CDouble
+           , Summation s (Identity Float)
+           )
+        => (forall a. Summation s a => s a -> a)
+        -> [TestTree]
+testSum f =
+  [ testGroup "Float" $ testSumOnType @Float f
+  , testGroup "Double" $ testSumOnType @Double f
+  , testGroup "CFloat" $ testSumOnType @CFloat f
+  , testGroup "CDouble" $ testSumOnType @CDouble f
+  , testGroup "Identity Float" $ testSumOnType @(Identity Float) f
+  ]
+
+testSumOnType :: forall a s. (Arbitrary a, Show a, Summation s a) => (s a -> a) -> [TestTree]
+testSumOnType f =
+  [ testProperty "t_sum" $ t_sum @a (sum f)
+  , testProperty "t_sum_error" $ t_sum_error @a (sum f)
+  , testProperty "t_sum_shifted" $ t_sum_shifted @a (sum f)
+  ]
+
+instance Arbitrary a => Arbitrary (KahanSum a) where
     arbitrary = toKahan <$> arbitrary
     shrink = map toKahan . shrink . fromKahan
 
-toKahan :: (Double, Double) -> KahanSum
+toKahan :: (a, a) -> KahanSum a
 toKahan (a,b) = KahanSum a b
 
-fromKahan :: KahanSum -> (Double, Double)
+fromKahan :: KahanSum a -> (a, a)
 fromKahan (KahanSum a b) = (a,b)
 
-instance Arbitrary KBNSum where
+instance Arbitrary a => Arbitrary (KBNSum a) where
     arbitrary = toKBN <$> arbitrary
     shrink = map toKBN . shrink . fromKBN
 
-toKBN :: (Double, Double) -> KBNSum
+toKBN :: (a, a) -> KBNSum a
 toKBN (a,b) = KBNSum a b
 
-fromKBN :: KBNSum -> (Double, Double)
+fromKBN :: KBNSum a -> (a, a)
 fromKBN (KBNSum a b) = (a,b)
 
-instance Arbitrary KB2Sum where
+instance Arbitrary a => Arbitrary (KB2Sum a) where
     arbitrary = toKB2 <$> arbitrary
     shrink = map toKB2 . shrink . fromKB2
 
-toKB2 :: (Double, Double, Double) -> KB2Sum
+toKB2 :: (a, a, a) -> KB2Sum a
 toKB2 (a,b,c) = KB2Sum a b c
 
-fromKB2 :: KB2Sum -> (Double, Double, Double)
+fromKB2 :: KB2Sum a -> (a, a, a)
 fromKB2 (KB2Sum a b c) = (a,b,c)