I'm trying to gauge interest in upstreaming some more generalized floating point compensated summations.
This is useful to me for use with the [`ad`](https://hackage.haskell.org/package/ad-4.4) library, where [`grad`](https://hackage.haskell.org/package/ad-4.4/docs/Numeric-AD.html#v:grad) takes a function `(Traversable f, Num a) => f (Reverse s a) -> Reverse s a)`. I use the `RealFloat` instance of `Reverse s a` to implement a loss function, which includes compensated floating point arithmetic. 

Currently [all tests pass](https://travis-ci.com/github/414owen/math-functions/builds/186735022) apart from GHCJS 8.4. I'll look into it if there's enough interest in merging this.

I've run some preliminary benchmarks, which seem very promising, and I think establish that any potential data boxing coming from this generalization won't do too much harm to a user.

# Old

```
$ cabal v2-run -O2 math-functions-bench

benchmarked Sum/naive
time                 2.783 ms   (2.747 ms .. 2.826 ms)
                     0.999 R²   (0.998 R² .. 1.000 R²)
mean                 2.704 ms   (2.691 ms .. 2.721 ms)
std dev              52.06 μs   (41.70 μs .. 69.10 μs)

benchmarked Sum/kahan
time                 5.854 ms   (5.827 ms .. 5.876 ms)
                     1.000 R²   (1.000 R² .. 1.000 R²)
mean                 5.862 ms   (5.849 ms .. 5.879 ms)
std dev              45.84 μs   (33.02 μs .. 62.22 μs)

benchmarked Sum/kbn
time                 6.279 ms   (6.252 ms .. 6.301 ms)
                     1.000 R²   (1.000 R² .. 1.000 R²)
mean                 6.277 ms   (6.268 ms .. 6.305 ms)
std dev              45.21 μs   (21.53 μs .. 86.52 μs)

benchmarked Sum/kb2
time                 4.389 ms   (4.367 ms .. 4.408 ms)
                     1.000 R²   (0.999 R² .. 1.000 R²)
mean                 4.400 ms   (4.386 ms .. 4.427 ms)
std dev              63.61 μs   (32.15 μs .. 108.3 μs)
```

# New

```
$ cabal v2-run -O2 math-functions-bench

benchmarked Sum/naive
time                 2.760 ms   (2.739 ms .. 2.781 ms)
                     1.000 R²   (1.000 R² .. 1.000 R²)
mean                 2.719 ms   (2.712 ms .. 2.728 ms)
std dev              27.64 μs   (22.99 μs .. 34.11 μs)

benchmarked Sum/kahan
time                 5.835 ms   (5.813 ms .. 5.855 ms)
                     1.000 R²   (1.000 R² .. 1.000 R²)
mean                 5.850 ms   (5.838 ms .. 5.867 ms)
std dev              42.72 μs   (26.45 μs .. 62.18 μs)

benchmarked Sum/kbn
time                 6.315 ms   (6.285 ms .. 6.343 ms)
                     1.000 R²   (1.000 R² .. 1.000 R²)
mean                 6.289 ms   (6.276 ms .. 6.307 ms)
std dev              45.34 μs   (33.86 μs .. 56.30 μs)

benchmarked Sum/kb2
time                 4.442 ms   (4.405 ms .. 4.476 ms)
                     1.000 R²   (0.999 R² .. 1.000 R²)
mean                 4.405 ms   (4.392 ms .. 4.427 ms)
std dev              51.35 μs   (37.47 μs .. 79.27 μs)
```