@414owen 2020-09-28
Okay, here are some [more complete benchmarks](https://gist.github.com/414owen/7d8b00a639a5f1d44f915a5f8dd1634b)

I also dumped the assembly for some minimal examples ([old](https://haskell.godbolt.org/z/ME1Ted), [new](https://haskell.godbolt.org/z/c5r6To)).

The new ASM doesn't look great, but I'm not really in a position to evaluate it.
The benchmarks don't seem to care too much...