@414owen 2020-09-28
I'm passing [`grad`](https://hackage.haskell.org/package/ad-4.4/docs/Numeric-AD.html#v:grad), and [`optimize`](https://hackage.haskell.org/package/nonlinear-optimization-ad-0.2.4/docs/Numeric-Optimization-Algorithms-HagerZhang05-AD.html#v:optimize) a function `(Traversable f, RealFloat a) => f a -> a`, where `a ~ Reverse s a`.
Elsewhere, I use the function where `a ~ Double`.