@414owen 2020-10-04
Here are some more easily comparable benchmarks: https://gist.github.com/414owen/ea366fc110a4e416ae9ceea035689a03

The new generalized version with `-O2` is slightly faster than the current version in every test.
I don't know why this is the case.