@Shimuuar 2020-10-04
I looked into core and yes GHC unboxes KBN/etc accumulators so inner loop has type `Double# -> Double# -> Int# -> Double`. It means it doesn't matter whether accumulator is boxed or not. On one hand it's pretty much ideal case for optimizer: simple loop where everything is inlined. On other it's more or less how accumulator is intended to be used. 

Question turns out to be how easy to break optimization above? I can't invent one on the spot so I need to think a bit about it. 

> The new ASM doesn't look great, but I'm not really in a position to evaluate it.

AFAIK NCG never was among best so it could generate suboptimal assembly

> I'm passing grad, and optimize a function ...

And how does KBN & friends enter the picture?